Moved IntegrationException into http and made it the superclass of RemoteIntegrationException.
