+++ config.py
@@ -5,7 +5,6 @@ import os
 import json
 import logging
 import copy
-from util import LanguageCodes
 from sqlalchemy.engine.url import make_url
 from flask_babel import lazy_gettext as _
 
@@ -14,34 +13,18 @@ from entrypoint import EntryPoint
 
 from sqlalchemy.exc import ArgumentError
 
-
-class IntegrationException(Exception):
-    """There was a problem in the setup or usage of an external integration.
-    """
-    def __init__(self, message, diagnostic=None):
-        """Constructor.
-
-        :param message: The normal message passed to any Exception
-        constructor.
-
-        :param diagnostic: An extra human-readable explanation of the
-        problem.  This may include instructions on what bits of the
-        integration configuration might need to be changed.
-
-        For example, an API key might be wrong, or the API key might
-        be correct but the API provider might not have granted that
-        key enough permissions.
-        """
-        super(IntegrationException, self).__init__(message)
-        self.diagnostic = diagnostic
+from util import LanguageCodes
+from util.http import IntegrationException
 
 
 class CannotLoadConfiguration(IntegrationException):
     """The current configuration of an external integration, or of the
     site as a whole, is in an incomplete or inconsistent state.
 
-    This is more specific than an IntegrationException because it assumes the
-    problem is evident just by looking at the current configuration.
+    This is more specific than a base IntegrationException because it
+    assumes the problem is evident just by looking at the current
+    configuration, with no need to actually talk to the foreign
+    server.
     """
     pass
